+++ packages/cspell-io/src/index.ts
@@ -3,6 +3,7 @@ export * from './common/index.js';
 export { compareStats, createTextFileResource } from './common/index.js';
 export type { CSpellIO } from './CSpellIO.js';
 export { CSpellIONode, getDefaultCSpellIO } from './CSpellIONode.js';
+export { createVirtualFS, getDefaultVirtualFs } from './CVirtualFS.js';
 export {
     getStat,
     getStatSync,
@@ -20,5 +21,4 @@ export { isFileURL, isUrlLike, toFileURL, toURL, urlBasename, urlDirname } from 
 export type { VFileSystem, VfsDirEntry, VfsStat } from './VFileSystem.js';
 export { FSCapabilityFlags } from './VFileSystem.js';
 export type { VFileSystemProvider, VirtualFS, VProviderFileSystem } from './VirtualFS.js';
-export { createVirtualFS, getDefaultVirtualFs } from './VirtualFS.js';
 export { createRedirectProvider } from './VirtualFS/redirectProvider.js';
